Initial commit of sdl-opengl-cpp project
This is a C++ wrapper for the SDL OpenGL functions.  It can be
compiled with and without exceptions.

It needs a lot of work to be usable by others, but it's a start to
abstracting out the OpenGL code.

Next versions will have simple initialization and uninitialization of
the library.