Move all the duplicate error checking methods into an Errors class
Also add a MoveChecker class that removes duplicate move copy /
assignment checking for valid but unspecified state.

This is the bulk of most of the refactor for simplifying errors.

I may add support for Boost LEAF.  But as it is now this provides
a much cleaner and DRY implementation of no exceptions code.