Refactor the library so initialization is clearner and more RAII.
Refactor the library so initialization is clearner and more RAII.
Add a SDL class and SDLWindow class.  This lets us mock out more of
the SDL library for testing on headless machines.

After this we'll add a better interface for initialization and the
rungl function.  std::function callbacks and more clean entry and exit
hooks.