Get some initial work in for SDL_ttf support.
The actual C++ SDL_ttf library is in a separate project and
repository.  This just includes necessary functions and features from
SDL such as some of the SDL_Surface and texture functions.

Needs more tests but this will let me reference a commit from
FetchContent to build out the SDL_ttf CMake files.